Move ProductCategoryController to Api namespace
